Add generated proto files as source dependencies
Adds generated proto files as source files and dependencies for the compilation of the iOS project. This allows for the generated files to be visible throughout the project as source files. These files are being generated in the pre-compilation phase in the build scheme to ensure these files are available to be used as dependent source files.

Signed-off-by: Shannyn Telander <telander@bitwise.io>