Refactor XOApi for better Swift practices
SwiftLint gives an error for the use of force casting, which seems to
be valid in the case where we are parsing data from a web service.

Instead of force casting, this has been changed to an optional cast
with a guard. This way we can handle an error caused by an unexpected
response more elegantly. The current print statements could be replaced
with some sort of UI alert in a future commit.

Also swaps the NS types for Swift native types and custom models. NS
types are used for Objective-C compatibility and it is not recommended
to use them over Swift native types.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>